+++ L25-di/homework-template/src/main/java/ru/otus/App.java
@@ -3,6 +3,7 @@ package ru.otus;
 import ru.otus.appcontainer.AppComponentsContainerImpl;
 import ru.otus.appcontainer.api.AppComponentsContainer;
 import ru.otus.config.AppConfig;
+import ru.otus.config.AppConfigExt;
 import ru.otus.services.GameProcessor;
 import ru.otus.services.GameProcessorImpl;
 
@@ -30,7 +31,7 @@ public class App {
         //AppComponentsContainer container = new AppComponentsContainerImpl("ru.otus.config");
 
         // Обязательный вариант
-        AppComponentsContainer container = new AppComponentsContainerImpl(AppConfig.class);
+        AppComponentsContainer container = new AppComponentsContainerImpl(AppConfig.class, AppConfigExt.class);
 
         // Приложение должно работать в каждом из указанных ниже вариантов
         GameProcessor gameProcessor = container.getAppComponent(GameProcessor.class);
