+++ L25-di/homework-template/src/main/java/ru/otus/config/AppConfig.java
@@ -12,21 +12,11 @@ public class AppConfig {
         return new EquationPreparerImpl();
     }
 
-    @AppComponent(order = 1, name = "playerService")
-    public PlayerService playerService(IOService ioService) {
-        return new PlayerServiceImpl(ioService);
-    }
-
     @AppComponent(order = 2, name = "gameProcessor")
     public GameProcessor gameProcessor(IOService ioService,
                                        PlayerService playerService,
                                        EquationPreparer equationPreparer) {
         return new GameProcessorImpl(ioService, equationPreparer, playerService);
-    }
-
-    @AppComponent(order = 0, name = "ioService")
-    public IOService ioService() {
-        return new IOServiceStreams(System.out, System.in);
     }
 
 }
