+++ L25-di/homework-template/src/main/java/ru/otus/config/AppConfigExt.java
@@ -1,0 +1,28 @@
+package ru.otus.config;
+
+import ru.otus.appcontainer.api.AppComponent;
+import ru.otus.appcontainer.api.AppComponentsContainerConfig;
+import ru.otus.services.EquationPreparer;
+import ru.otus.services.EquationPreparerImpl;
+import ru.otus.services.GameProcessor;
+import ru.otus.services.GameProcessorImpl;
+import ru.otus.services.IOService;
+import ru.otus.services.IOServiceStreams;
+import ru.otus.services.PlayerService;
+import ru.otus.services.PlayerServiceImpl;
+
+@AppComponentsContainerConfig(order = 1)
+public class AppConfigExt {
+
+    @AppComponent(order = 1, name = "playerService")
+    public PlayerService playerService(IOService ioService) {
+        return new PlayerServiceImpl(ioService);
+    }
+
+
+    @AppComponent(order = 0, name = "ioService")
+    public IOService ioService() {
+        return new IOServiceStreams(System.out, System.in);
+    }
+
+}
